+++ pytsite/cleanup/__init__.py
@@ -1,10 +1,10 @@
 """PytSite Cleanup Module.
 """
 from pytsite import events as _events
-from ._eh import cron_daily
+from ._eh import cron_hourly
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
-_events.listen('pytsite.cron.daily', cron_daily)
+_events.listen('pytsite.cron.hourly', cron_hourly)