+++ pytsite/geo_ip/_error.py
@@ -1,0 +1,9 @@
+"""PytSite GeoIP Errors.
+"""
+__author__ = 'Alexander Shepetko'
+__email__ = 'a@shepetko.com'
+__license__ = 'MIT'
+
+
+class ResolveError(Exception):
+    pass